pyproject.toml
[project]
name = "doxearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf>=6.6.2",
    "python-docx>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
